[apirambo] rename MemoryAccessBase.h as MemoryAccessHelpers.h
