+++ epochX/cudacpp/ee_mumu/SubProcesses/P1_Sigma_sm_epem_mupmum/MemoryAccessMomenta.h
@@ -5,7 +5,7 @@
 #include "mgOnGpuTypes.h"
 //#include "mgOnGpuVectors.h"
 
-#include "MemoryAccessBase.h"
+#include "MemoryAccessHelpers.h"
 
 //----------------------------------------------------------------------------
